refactor: Clean up unused imports and variables per RUST_CODING_STANDARDS
Applied clippy automatic fixes to remove:
- Unused imports across all modules
- Unused variables (prefixed with _ where needed)
- Fixed simd_json mutable slice parameter
- Cleaned up unnecessary mut qualifiers

Files modified:
- capnweb-core/src/protocol/* (multiple modules)
- capnweb-core/src/codec.rs (fixed simd_json signature)
- capnweb-core/src/promise.rs

Testing:
✅ Rust build successful
✅ TypeScript client tests passing

Following RUST_CODING_STANDARDS.md guidelines for code clarity.

+++ capnweb-core/src/protocol/message.rs
@@ -53,7 +53,7 @@ impl Message {
                     return Err(MessageError::InvalidPull);
                 }
                 let import_id = arr[1].as_i64()
-                    .ok_or(MessageError::InvalidImportId)? as i64;
+                    .ok_or(MessageError::InvalidImportId)?;
                 Ok(Message::Pull(ImportId(import_id)))
             }
 
@@ -62,7 +62,7 @@ impl Message {
                     return Err(MessageError::InvalidResolve);
                 }
                 let export_id = arr[1].as_i64()
-                    .ok_or(MessageError::InvalidExportId)? as i64;
+                    .ok_or(MessageError::InvalidExportId)?;
                 let expr = Expression::from_json(&arr[2])?;
                 Ok(Message::Resolve(ExportId(export_id), expr))
             }
@@ -72,7 +72,7 @@ impl Message {
                     return Err(MessageError::InvalidReject);
                 }
                 let export_id = arr[1].as_i64()
-                    .ok_or(MessageError::InvalidExportId)? as i64;
+                    .ok_or(MessageError::InvalidExportId)?;
                 let expr = Expression::from_json(&arr[2])?;
                 Ok(Message::Reject(ExportId(export_id), expr))
             }
@@ -82,7 +82,7 @@ impl Message {
                     return Err(MessageError::InvalidRelease);
                 }
                 let import_id = arr[1].as_i64()
-                    .ok_or(MessageError::InvalidImportId)? as i64;
+                    .ok_or(MessageError::InvalidImportId)?;
                 let refcount = arr[2].as_u64()
                     .ok_or(MessageError::InvalidRefcount)? as u32;
                 Ok(Message::Release(ImportId(import_id), refcount))
